requirements.txt
Flask
pywebview
configparser
pillow
torch
torchvision
torchaudio